add some coding documentation
