+++ src/helpers.ts
@@ -1,19 +1,28 @@
+/**
+ * Home Assistant Helper Functions
+ *
+ * A collection of utility functions for working with Home Assistant entities,
+ * states, and configurations. These functions handle entity management,
+ * state retrieval, and UI element creation.
+ */
+
 import { type TemplateResult, html } from 'lit';
 
-import { actionHandler } from './action-handler';
-import { handleClickAction } from './handle-action';
+import { actionHandler, handleClickAction } from './action-handler';
 import { getClimateStyles, getEntityIconStyles } from './styles';
-import type {
-  Area,
-  Config,
-  Device,
-  Entity,
-  EntityConfig,
-  EntityInformation,
-  HomeAssistant,
-  State,
-} from './types';
+import type { EntityInformation, Config, EntityConfig } from './types/config';
+import type { HomeAssistant, State, Entity, Device, Area } from './types/homeassistant';
 
+
+/**
+ * Creates a state icon element for an entity
+ *
+ * @param {HTMLElement} element - The parent element that will contain the icon
+ * @param {HomeAssistant} hass - The Home Assistant instance
+ * @param {EntityInformation} entity - Information about the entity
+ * @param {String[]} classes - CSS classes to apply to the icon container
+ * @returns {TemplateResult} A Lit template containing the icon element
+ */
 export const createStateIcon = (
   element: HTMLElement,
   hass: HomeAssistant,
@@ -40,30 +49,72 @@ export const createStateIcon = (
   </div>`;
 };
 
+/**
+ * Retrieves the state of an entity
+ *
+ * @param {HomeAssistant} hass - The Home Assistant instance
+ * @param {string} [entityId] - The ID of the entity
+ * @param {boolean} [fakeState=false] - Whether to create a fake state if none exists
+ * @returns {State | undefined} The entity's state or undefined
+ */
 export const getState = (
   hass: HomeAssistant,
   entityId?: string,
   fakeState: boolean = false,
 ): State | undefined => {
   if (!entityId) return undefined;
+
   const state =
     (hass.states as { [key: string]: any })[entityId] ||
     (fakeState ? { entity_id: entityId } : undefined);
+
   if (!state) return undefined;
-  return { ...state, getDomain: () => state.entity_id.split('.')[0] };
+
+  return {
+    ...state,
+    getDomain: () => state.entity_id.split('.')[0],
+  };
 };
 
+/**
+ * Retrieves entity information
+ *
+ * @param {HomeAssistant} hass - The Home Assistant instance
+ * @param {string} entityId - The ID of the entity
+ * @returns {Entity} The entity information
+ */
 export const getEntity = (hass: HomeAssistant, entityId: string): Entity =>
   (hass.entities as { [key: string]: any })[entityId];
 
+/**
+ * Retrieves device information
+ *
+ * @param {HomeAssistant} hass - The Home Assistant instance
+ * @param {string} deviceId - The ID of the device
+ * @returns {Device} The device information
+ */
 export const getDevice = (hass: HomeAssistant, deviceId: string): Device =>
   (hass.devices as { [key: string]: any })[deviceId];
 
+/**
+ * Retrieves area information
+ *
+ * @param {HomeAssistant} hass - The Home Assistant instance
+ * @param {string} deviceId - The ID of the device
+ * @returns {Area | undefined} The area information
+ */
 export const getArea = (
   hass: HomeAssistant,
   deviceId: string,
 ): Area | undefined => (hass.areas as { [key: string]: any })[deviceId];
 
+/**
+ * Gets entities with problems in a specific area
+ *
+ * @param {HomeAssistant} hass - The Home Assistant instance
+ * @param {string} area - The area to check for problems
+ * @returns {Object} Object containing problem entities and existence flag
+ */
 export const getProblemEntities = (
   hass: HomeAssistant,
   area: string,
@@ -71,6 +122,7 @@ export const getProblemEntities = (
   problemEntities: string[];
   problemExists: boolean;
 } => {
+  // Find entities labeled as problems in the specified area
   const problemEntities = Object.keys(hass.entities).filter((entityId) => {
     const entity = hass.entities[entityId];
     if (!entity?.labels?.includes('problem')) return false;
@@ -79,6 +131,7 @@ export const getProblemEntities = (
     return [entity.area_id, device?.area_id].includes(area);
   });
 
+  // Check if any problem entities are currently active
   const problemExists = problemEntities.some((entityId) => {
     const entityState = hass.states[entityId];
     if (!entityState) return false;
@@ -91,7 +144,18 @@ export const getProblemEntities = (
   };
 };
 
-export const getIconEntities = (hass: HomeAssistant, config: Config) => {
+/**
+ * Gets entities to display icons for
+ *
+ * @param {HomeAssistant} hass - The Home Assistant instance
+ * @param {Config} config - The configuration object
+ * @returns {EntityInformation[]} Array of configured entities
+ */
+export const getIconEntities = (
+  hass: HomeAssistant,
+  config: Config,
+): EntityInformation[] => {
+  // Define base entities for the area
   const baseEntities = [
     `light.${config.area}_light`,
     `switch.${config.area}_fan`,
@@ -99,24 +163,28 @@ export const getIconEntities = (hass: HomeAssistant, config: Config) => {
 
   const configEntities = config.entities || [];
 
+  // Combine base and config entities unless fan is removed
   const entities = config.remove_fan
     ? configEntities
     : baseEntities.concat(configEntities);
 
+  // Process and transform entities
   const states = entities
     .map((entity) => {
-      // handle the convenience string format
+      // Transform string format to entity config
       if (typeof entity === 'string') {
         entity = { entity_id: entity };
       }
 
       const state = getState(hass, entity.entity_id);
       if (!state) return undefined;
+
       const useClimateColors =
         !config.skip_climate_colors && state.getDomain() === 'climate';
 
       const { climateStyles, climateIcons } = getClimateStyles();
 
+      // Create entity information with defaults and climate handling
       return {
         config: {
           tap_action: { action: 'toggle' },
@@ -135,45 +203,61 @@ export const getIconEntities = (hass: HomeAssistant, config: Config) => {
         },
       } as EntityInformation;
     })
-    .filter((entity) => entity !== undefined);
+    .filter((entity): entity is EntityInformation => entity !== undefined);
+
   return states;
 };
 
-export const getRoomEntity = (hass: HomeAssistant, config: Config) => {
+/**
+ * Gets the room entity information
+ *
+ * @param {HomeAssistant} hass - The Home Assistant instance
+ * @param {Config} config - The configuration object
+ * @returns {EntityInformation} The room entity information
+ */
+export const getRoomEntity = (
+  hass: HomeAssistant,
+  config: Config,
+): EntityInformation => {
   const roomEntityId = `light.${config.area}_light`;
-  const roomEntity: EntityInformation | undefined = config.entity
-    ? typeof config.entity === 'string'
-      ? // convenience transform - use their entity
-        {
-          config: {
-            entity_id: config.entity,
-            hold_action: { action: 'more-info' },
-            double_tap_action: { action: 'none' },
-          },
-          state: getState(hass, config.entity),
-        }
-      : {
-          config: {
-            hold_action: { action: 'more-info' },
-            double_tap_action: { action: 'none' },
-            ...config.entity,
-          },
-          state: getState(hass, config.entity.entity_id),
-        }
-    : // else use the room light
-      {
+
+  // Handle different entity configuration formats
+  if (config.entity) {
+    if (typeof config.entity === 'string') {
+      // String format
+      return {
         config: {
-          entity_id: roomEntityId,
-          icon: getArea(hass, config.area)?.icon,
-          tap_action: {
-            action: 'navigate',
-            navigation_path: config.navigate ?? config.area.replace('_', '-'),
-          },
+          entity_id: config.entity,
           hold_action: { action: 'more-info' },
           double_tap_action: { action: 'none' },
-        } as EntityConfig,
-        // fake state in case room entity doesn't exist
-        state: getState(hass, roomEntityId, true),
+        },
+        state: getState(hass, config.entity),
       };
-  return roomEntity;
+    } else {
+      // Object format
+      return {
+        config: {
+          hold_action: { action: 'more-info' },
+          double_tap_action: { action: 'none' },
+          ...config.entity,
+        },
+        state: getState(hass, config.entity.entity_id),
+      };
+    }
+  }
+
+  // Default room light configuration
+  return {
+    config: {
+      entity_id: roomEntityId,
+      icon: getArea(hass, config.area)?.icon,
+      tap_action: {
+        action: 'navigate',
+        navigation_path: config.navigate ?? config.area.replace('_', '-'),
+      },
+      hold_action: { action: 'more-info' },
+      double_tap_action: { action: 'none' },
+    } as EntityConfig,
+    state: getState(hass, roomEntityId, true),
+  };
 };
